requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pandas
sqlalchemy
python-dotenv
jinja2
gunicorn
